djangoapp/requirements.txt
Django>=5.1.6,<5.2
psycopg2-binary>=2.9.10,<2.10
djangorestframework>=3.15.2,<3.16
paho-mqtt>=2.1.0,<2.2
python-dotenv>=1.0.1,<1.1